+++ src/node/components/draw/draw.ts
@@ -12,23 +12,16 @@ export class NodeDrawComponent implements IComponent {
 
     public Update(deltaTime: number): void {
         this.Clear()
+        this.Entity.AdjustSize()
         this.Draw()
     }
 
     private Draw(): void {
-        if (this.Entity.Image) {
-            this.Canvas.RenderImage(
-                this.Entity.Start,
-                this.Entity.Size,
-                this.Entity.Image
-            )
-        } else {
-            this.Canvas.FillRect(
-                this.Entity.Start,
-                this.Entity.Size,
-                this.Entity.color
-            )
-        }
+        this.Canvas.FillRect(
+            this.Entity.Start,
+            this.Entity.Size,
+            this.Entity.color
+        )
     }
 
     public Clear(): void {