+++ src/snake/assets.ts
@@ -1,15 +1,7 @@
-import headDown from '@/images/snake_head_down.png'
-import headUp from '@/images/snake_head_up.png'
-import headRight from '@/images/snake_head_right.png'
-import headLeft from '@/images/snake_head_left.png'
+import headSrc from '@/images/snake_head_right.png'
 import tailSrc from '@/images/snake_part.png'
 
 export const SNAKE = {
-    HEAD: {
-        DOWN: headDown,
-        UP: headUp,
-        RIGHT: headRight,
-        LEFT: headLeft,
-    },
+    HEAD: headSrc,
     TAIL: tailSrc,
 }