+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
+## 0.0.2 (2021-08-15)
+### Improvements
+ - Improved snake movement rendering
+ - Entities and Components rewrite for better usage of ECS
+### Features
+ - Added a `Save` and `Load` event into `Game` (you can use the console with the `game` object to emit the events `Save` and `Load`)
+
 ## 0.0.1 (2021-08-05)
 
 ### Features
  - Snake Entity
  - Keyboard Usage (You can use the `arrows` or `WASD` to move the snake, and `Enter` to pause or unpause the game)
- - Event Handling (you can use the console using the `game` object to emit events such as `Pause` and `Start` or `Restart`)
+ - Event Handling (you can use the console with the `game` object to emit events such as `Pause` and `Start` or `Restart`)
  - Configuration Control (you can use the console by modifying the available attributes from the `gameSettings` object) 
  - Food and Basic Interface presentation